Fixed many test bugs and simplified test suite.
